pyproject.toml
[tool.poetry]
name = "test-teamup"
version = "0.1.0"
description = ""
authors = ["Andrei Rostovtsev <105871945+Sprin94@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "test_teamup"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
black = "^23.3.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
